+++ Realm/RLMSyncCredentials.h
@@ -48,6 +48,12 @@ extern RLMIdentityProvider const RLMIdentityProviderCloudKit;
 /// A JSON Web Token as an identity provider.
 extern RLMIdentityProvider const RLMIdentityProviderJWT;
 
+/// An Anonymous account as an identity provider.
+extern RLMIdentityProvider const RLMIdentityProviderAnonymous;
+
+/// A Nickname account as an identity provider.
+extern RLMIdentityProvider const RLMIdentityProviderNickname;
+
 /**
  Opaque credentials representing a specific Realm Object Server user.
  */
@@ -88,6 +94,16 @@ extern RLMIdentityProvider const RLMIdentityProviderJWT;
  Construct and return credentials from a JSON Web Token.
  */
 + (instancetype)credentialsWithJWT:(NSString *)token;
+
+/**
+ Construct and return anonymous credentials
+ */
++ (instancetype)anonymousCredentials;
+    
+/**
+ Construct and return credentials from a nickname
+ */
++ (instancetype)credentialsWithNickname:(NSString *)nickname isAdmin:(BOOL)isAdmin;
 
 /**
  Construct and return special credentials representing a token that can